Change developer profile type directory
