+++ src/containers/DeveloperProfile.tsx
@@ -1,6 +1,25 @@
 import React, { createContext, PropsWithChildren, useContext } from 'react';
 
-import { DeveloperProfileGQL } from '../components/DetailsCard/types';
+export type DeveloperProfileGQL = {
+  last_name: string;
+  is_freelancer: boolean;
+  first_name: string;
+  email: string;
+  phone: string;
+  country: string;
+  city: string;
+  avatar: {
+    publicURL: string;
+  };
+  social_media: {
+    facebook: string;
+    github: string;
+    instagram: string;
+    twitter: string;
+  };
+  position: string;
+  cv: string;
+};
 
 type DeveloperProfile = DeveloperProfileGQL | null;
